requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.3
django-ckeditor==6.2.0
django-heroku==0.3.1
django-js-asset==2.0.0
gunicorn==20.1.0
Pillow==9.0.1
psycopg2==2.9.3
python-decouple==3.6
python-dotenv==0.19.2
sqlparse==0.4.2
tzdata==2021.5
whitenoise==6.0.0
